requirements.txt
Django==3.2
psycopg2-binary==2.9.1
python-dotenv>=1.0.0
requests>=2.28.0
SQLAlchemy>=1.4,<2
# Any other dependencies